requirements.txt
celery
django-celery-results
asgiref==3.9.1
sqlparse==0.5.3
typing_extensions==4.14.1
psycopg2-binary
django-redis
requests
drf-yasg
